modify: 言語選択、切り替え処理を別コンポーネントに移管
