+++ src/components/Beginning.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
     import Introduction from "./Introduction.svelte";
     import BottomArrow from "./BottomArrow.svelte";
+    import SelectLang from "./SelectLang.svelte";
     import { lang } from "../scripts/lang";
     const message_jp =
         `ここは日本在住のエンジニア、 \nMIHARA Keishi のポートフォリオサイトです。`.replace(
@@ -12,30 +13,12 @@
         /\n/g,
         "<br />"
     );
-
-    let langKey: string;
-
-    lang.subscribe((value) => {
-        langKey = value;
-    });
-
-    const switchLanguage = (l: string) => {
-        return () => {
-            lang.update((_) => {
-                return l;
-            });
-        };
-    };
 </script>
 
 <div class="beginning">
     <h4>{@html message_jp}</h4>
     <h5>{@html message_en}</h5>
-    {#if langKey == "ja"}
-        <a href={null} on:click={switchLanguage("en")}>English here.</a>
-    {:else}
-        <a href={null} on:click={switchLanguage("ja")}>日本語はこちら</a>
-    {/if}
+    <SelectLang />
     <Introduction />
     <BottomArrow />
 </div>
@@ -44,10 +27,6 @@
     h4,
     h5 {
         font-family: "Noto Sans JP";
-    }
-    a {
-        cursor: pointer;
-        text-decoration: underline;
     }
     .beginning {
         background-color: #110011;
